refactor(app): extract App component and tighten types; improve worker spawn options
