+++ packages/app/src/renderer/global.d.ts
@@ -3,5 +3,3 @@ declare global {
     aideon: { version: string };
   }
 }
-export {};
-
